requirements.txt
Django==3.2.7
django-environ==0.7.0
factory-boy==3.2.0
djangorestframework==3.12.4